back-end/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pipenv==2022.9.4
platformdirs==2.5.2
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
twilio==7.14.1
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==2.2.2
